UPD: update tests for new objective function and shifted time periods.
Time elapsed is now calculated by the number of repairs allowed in each
repair window.  The objective function is muiltiplied by time_elapsed
for each network.  Time periods are alos now indexed from 0.  The 0
network is prior to any repairs, and network 1 is the first period when
repairs are completed and devices enabled.

Time_elapsed for network 0 is based on the number of repairs e.g. if 10
repairs are to be performed, the network is its fully damaged state for
10 hours.  then in network 1, the 10 devices are enabled and a new
repair period starts.