New objective - multiply by time_elapsed (#2)
* UPD: time elapsed adjusted based on the number of repairs in a period
* UPD: time_elapsed/repairs per period now account for the total repairs.

Previously, each time period is exactly equal, and rounded up to
provide sufficient repairs options to fix the enetire network.  Now, the
final time period is calculated to match the total repairs needed, e.g.
for 25 repairs in 3 periods the time_elapsed/repair_per_period is
[10,10,5] rather than [10,10,10].  This change allows for better
comparison of restorations with various periods.

* UPD: update tests for new objective function and shifted time periods.

Time elapsed is now calculated by the number of repairs allowed in each
repair window.  The objective function is multiplied by time_elapsed
for each network.  Time periods are also now indexed from 0.  The 0
network is prior to any repairs, and network 1 is the first period when
repairs are completed and devices enabled.

Time_elapsed for network 0 is based on the number of repairs e.g. if 10
repairs are to be performed, the network is its fully damaged state for
10 hours.  then in network 1, the 10 devices are enabled and a new
repair period starts.

Objective function uses a smaller multiplier for gen, storage, etc. This
change ensures all loads are enabled, instead of focusing the objective
on generation.

name of clean_solution! now uses exclamation according to standard
naming.

* UPD: remove unstable tests

* UPD: comment out unused/undeveloped functions

* Update project manifest

* UPD: move repair/time_elapsed calculation to network build stage

* UPD: consolidate rop and rop_uc into rop