Update restoration_simulation to avoid mixed integers
restoration_simulation used to remove the damage indicator and
use run_rop to redispatch the generation.  This process fixed the
restoration decisions and allowed the redispatch to use a solver
like Ipopt because the mixed integer variables are fixed.

However, the complementary constraint on battery storage still used
a mixed integer.

This update creates a seperate subproblem for restoration simulation.
It is a multi-period opf that maxmizes load delivered, but can shed
load and allows (penalized) voltage violations.

Additional fixes:
added manifest.toml to. gitignore

changed name of solution_rop to solution_rop!

changed name of solution_mrsp to solution_mrsp!

added function for voltage violation on non-damaged buses
changed name of function constraint_bus_voltage to
constraint_bus_voltage_violation_damage to better represent
what the constraint does

changed _clean_status! to set status to the inactive value
if it is approximately that value, instead of rounding. This
prevents a load status of a sheddable load from being rounded
to status=0, and then being removed from the network when passed
to a subsequent problem (like simulation).

the test for restoration_simulation used update_data! instead of
update_status!.  This can cause errors where the reactive power of a
load is removed (set to NaN) by a DC power model for rop, then is
set to 0 when processed by  restoration_simulation. Instead, just the
status is updated from a solution to the network data for running
the simulation.