FIX: error when load exceeds generation
The constraint restore_all_items required all items to have
a z_* variable value of 1.  This also requires that loads/shunts
are fully powered in the final time-step. If there is not
enough generation to power the loads (some generation is status=0,
unrepairable) then the problem is infeasible.

This changed a solution to a rop test where two branches switched the
order in which they were repaired.