Fix incorect constraint
This constraint used the storage id `i` to index a bus when
constrainting storage to be inactive before bus. For example,
Storage 1 cannot be activated before its connected Bus 2 is activate.
In this case Bus 1 was referenced for storage 1.

The correct constraint implementation was included further down in the
function, but the incorrect constraint was never removed.