Bugfix: 44 (#45)
* Bugfix: 44

Fixes #44

Must not be covered in the tests, I will look into why.

* Rename iterative restore test file.  Add damaged bus.

* add storage element to iterative case

* test function load_power notification

function should notify if nw_id does not exist in network.  Was an issue
because I did not know why network was returning 0 load.

* Update language for run_iterative description

* Update iterative restoration tests.

Added damaged bus and storage to case study.
Remove the order of repaired components because of degeneracy.