Fix issues with total damage restoration using iterative (#55)
* Add total-damage scenario for iter testing

Test case does not include "time_elapsed" which also tests data handling

* Fix for hanlding time_elapsed in iterative

* Fix MLD objective function when no loads present

when no loads are present, maximum([]) of an empty
array is called, resulting in an error.  Update ensures
maximum is never called on an empty array.

* Change 3-bus iter case to be ROP feasible

* Add data processing if MLD is an empty problem

If iteration restoration is done on a completely damaged network, the
initial MLD run is an empty problem and no optimziation is done.  Added
data verification to add component dictionaries if  MLD is not run.